qlib-trading-platform/qlib-trading-backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.3

# Deep Learning (optional)
torch==2.1.1
torchvision==0.16.1

# Data Sources
yfinance==0.2.28
akshare==1.11.0
tushare==1.3.0

# Qlib Integration (if available)
# qlib==0.9.3  # Uncomment if qlib is available

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# File Processing
openpyxl==3.1.2
xlrd==2.0.1

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# API Documentation
fastapi-pagination==0.12.12

# Time Series Analysis
statsmodels==0.14.0

# Financial Data Processing
empyrical==0.5